Code Quality: fixed many checkstyle, pmd and findbugs warnings. Generally should have no effect on functionality except where division now employs rounding instead of trunction

+++ mapsforge-core/src/main/java/org/mapsforge/core/model/Dimension.java
@@ -39,7 +39,7 @@ public class Dimension implements Serializable {
 	 * @return the center point
 	 */
 	public Point getCenter() {
-		return new Point(this.width / 2, this.height / 2);
+		return new Point((float) this.width / 2, (float) this.height / 2);
 	}
 
 	@Override
